pyproject.toml
[project]
name = "django-ninja"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "guilherme", email = "vinespedro55@gmail.com"},
]
dependencies = [
    "django-ninja>=1.3.0",
    "pydantic>=2.9.1",
    "pandas>=2.2.2",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
